+++ lib/kitchen/directions/bake_chapter_solutions/main.rb
@@ -3,8 +3,13 @@
 module Kitchen
   module Directions
     module BakeChapterSolutions
-      def self.v1(chapter:, metadata_source:, uuid_prefix: '')
-        V1.new.bake(chapter: chapter, metadata_source: metadata_source, uuid_prefix: uuid_prefix)
+      def self.v1(chapter:, metadata_source:, uuid_prefix: '', classes: %w[free-response])
+        V1.new.bake(
+          chapter: chapter,
+          metadata_source: metadata_source,
+          uuid_prefix: uuid_prefix,
+          classes: classes
+        )
       end
     end
   end