+++ spec/directions/bake_chapter_solutions/main_spec.rb
@@ -5,7 +5,7 @@ require 'spec_helper'
 RSpec.describe Kitchen::Directions::BakeChapterSolutions do
   it 'calls v1' do
     expect_any_instance_of(Kitchen::Directions::BakeChapterSolutions::V1).to receive(:bake)
-      .with(chapter: 'chapter1', metadata_source: 'metadata', uuid_prefix: 'uuid')
-    described_class.v1(chapter: 'chapter1', metadata_source: 'metadata', uuid_prefix: 'uuid')
+      .with(chapter: 'chapter1', metadata_source: 'metadata', uuid_prefix: 'uuid', classes: 'class')
+    described_class.v1(chapter: 'chapter1', metadata_source: 'metadata', uuid_prefix: 'uuid', classes: 'class')
   end
 end