+++ spec/directions/bake_chapter_solutions/v1_spec.rb
@@ -26,6 +26,15 @@ RSpec.describe Kitchen::Directions::BakeChapterSolutions::V1 do
                 </div>
               </div>
             </section>
+            <section data-depth="1" id="0" class="practice">
+              <h3 data-type="title">PRACTICE</h3>
+              <div data-type="exercise">
+                <div data-type="problem">What is the airspeed velocity of an unladen swallow?</div>
+                <div data-type="solution" id='2'>
+                  <p>Something that Numbered Notes will handle</p>
+                </div>
+              </div>
+            </section>
           </div>
         </div>
       HTML
@@ -33,7 +42,10 @@ RSpec.describe Kitchen::Directions::BakeChapterSolutions::V1 do
   end
 
   it 'works' do
-    described_class.new.bake(chapter: chapter, metadata_source: metadata_element, uuid_prefix: '')
+    described_class.new.bake(chapter: chapter,
+                             metadata_source: metadata_element,
+                             uuid_prefix: '',
+                             classes: %w[free-response practice])
     expect(chapter).to match_normalized_html(
       <<~HTML
         <div data-type="chapter">
@@ -43,6 +55,12 @@ RSpec.describe Kitchen::Directions::BakeChapterSolutions::V1 do
               <h3 data-type="title">HOMEWORK</h3>
               <div data-type="exercise">
                 <div data-type="problem">How much wood does a wood chuck chuck?</div>
+              </div>
+            </section>
+            <section class="practice" data-depth="1" id="0">
+              <h3 data-type="title">PRACTICE</h3>
+              <div data-type="exercise">
+                <div data-type="problem">What is the airspeed velocity of an unladen swallow?</div>
               </div>
             </section>
           </div>
@@ -60,6 +78,9 @@ RSpec.describe Kitchen::Directions::BakeChapterSolutions::V1 do
             </div>
             <div data-type="solution" id="1">
               <p>Not much here because bakeNumberedNotes does that for us </p>
+            </div>
+            <div data-type="solution" id="2">
+              <p>Something that Numbered Notes will handle</p>
             </div>
           </div>
         </div>